+++ bluebao/Boye/消息通知管理/BoyeGoalLocaNotify.m
@@ -23,7 +23,7 @@
         return;
     }
     
-    [[ LocalNotify sharedNotify]  fireNotification:key At:model.fireDate  WithContent:[NSString stringWithFormat:@" %@到运动时间了，目标 %ld ",model.date_time,model.target] HasInterval:NSCalendarUnitWeekday];
+    [[ LocalNotify sharedNotify]  fireNotification:key At:model.fireDate  WithContent:[NSString stringWithFormat:@" %@到运动时间了，目标 %ld ",model.date_time,(long)model.target] HasInterval:NSCalendarUnitWeekday];
 }
 
 //取消通知
@@ -71,7 +71,7 @@
  **/
 + (NSString *) getFullNotifyKey:(NSInteger ) uid goalID:(NSInteger)goalID{
     
-    return [NSString stringWithFormat:@"%ld_%ld",uid,goalID];
+    return [NSString stringWithFormat:@"%ld_%ld",(long)uid,(long)goalID];
 }
